requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-debug-toolbar==4.1.0
black[d]==23.3.0
psycopg2==2.9.6
geopy==2.3.0
